# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deadline Guardian AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management
  - name: ai
    description: AI-powered features

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /tasks/dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [tasks]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /tasks/analytics/weekly:
    get:
      operationId: getWeeklyAnalytics
      tags: [tasks]
      summary: Get weekly productivity analytics
      responses:
        "200":
          description: Weekly analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyData"

  /ai/breakdown:
    post:
      operationId: generateBreakdown
      tags: [ai]
      summary: Generate AI task breakdown
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakdownRequest"
      responses:
        "200":
          description: Task breakdown with subtasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakdownResult"

  /ai/risk:
    post:
      operationId: analyzeRisk
      tags: [ai]
      summary: Analyze deadline risk for a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskRequest"
      responses:
        "200":
          description: Risk analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskResult"

  /ai/rescue:
    post:
      operationId: generateRescuePlan
      tags: [ai]
      summary: Generate emergency rescue plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescueRequest"
      responses:
        "200":
          description: Rescue plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RescueResult"

  /ai/coach:
    post:
      operationId: generateCoachInsights
      tags: [ai]
      summary: Generate AI coach report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachRequest"
      responses:
        "200":
          description: Coach insights report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      required: [id, title, deadline, priority, status, progress, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        deadline:
          type: string
        estimatedHours:
          type: ["number", "null"]
        availableHours:
          type: ["number", "null"]
        energyLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [todo, in_progress, completed]
        riskScore:
          type: ["integer", "null"]
        riskLevel:
          type: ["string", "null"]
          enum: [low, medium, high, critical, null]
        subtasks:
          type: ["string", "null"]
        riskAnalysis:
          type: ["string", "null"]
        rescuePlan:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    TaskInput:
      type: object
      required: [title, deadline, priority]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        deadline:
          type: string
        estimatedHours:
          type: number
        availableHours:
          type: number
        energyLevel:
          type: string
          enum: [low, medium, high]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [todo, in_progress, completed]

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        deadline:
          type: string
        estimatedHours:
          type: ["number", "null"]
        availableHours:
          type: ["number", "null"]
        energyLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [todo, in_progress, completed]
        riskScore:
          type: ["integer", "null"]
        riskLevel:
          type: ["string", "null"]
        subtasks:
          type: ["string", "null"]
        riskAnalysis:
          type: ["string", "null"]
        rescuePlan:
          type: ["string", "null"]

    DashboardSummary:
      type: object
      required: [totalTasks, completedTasks, highRiskTasks, productivityScore, upcomingDeadlines, recentActivity]
      properties:
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        highRiskTasks:
          type: integer
        productivityScore:
          type: number
        upcomingDeadlines:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        completionRate:
          type: number
        averageRisk:
          type: number

    WeeklyData:
      type: object
      required: [day, completed, created, riskAvg]
      properties:
        day:
          type: string
        completed:
          type: integer
        created:
          type: integer
        riskAvg:
          type: number

    BreakdownRequest:
      type: object
      required: [taskId, title]
      properties:
        taskId:
          type: integer
        title:
          type: string
        description:
          type: string
        estimatedHours:
          type: number
        deadline:
          type: string

    BreakdownResult:
      type: object
      required: [taskId, subtasks]
      properties:
        taskId:
          type: integer
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/Subtask"

    Subtask:
      type: object
      required: [title, estimatedTime, difficulty]
      properties:
        title:
          type: string
        estimatedTime:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        dependencies:
          type: array
          items:
            type: string
        completed:
          type: boolean

    RiskRequest:
      type: object
      required: [taskId, title, deadline, progress]
      properties:
        taskId:
          type: integer
        title:
          type: string
        description:
          type: string
        deadline:
          type: string
        progress:
          type: integer
        estimatedHours:
          type: number
        availableHours:
          type: number
        energyLevel:
          type: string

    RiskResult:
      type: object
      required: [taskId, riskScore, riskLevel, explanation, recommendations]
      properties:
        taskId:
          type: integer
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        explanation:
          type: string
        recommendations:
          type: array
          items:
            type: string

    RescueRequest:
      type: object
      required: [taskId, title, deadline, progress, riskScore]
      properties:
        taskId:
          type: integer
        title:
          type: string
        description:
          type: string
        deadline:
          type: string
        progress:
          type: integer
        estimatedHours:
          type: number
        availableHours:
          type: number
        riskScore:
          type: integer

    RescueResult:
      type: object
      required: [taskId, criticalTasks, tasksToSkip, scopeReduction, timeline, recoveryStrategy, recommendedTools]
      properties:
        taskId:
          type: integer
        criticalTasks:
          type: array
          items:
            type: string
        tasksToSkip:
          type: array
          items:
            type: string
        scopeReduction:
          type: string
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineItem"
        recoveryStrategy:
          type: string
        recommendedTools:
          type: array
          items:
            type: string

    TimelineItem:
      type: object
      required: [time, action, priority]
      properties:
        time:
          type: string
        action:
          type: string
        priority:
          type: string
          enum: [critical, high, medium, low]

    CoachRequest:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"

    CoachResult:
      type: object
      required: [assessment, feedback, recommendations, recoveryActions, motivationalMessage, productivityScore]
      properties:
        assessment:
          type: string
        feedback:
          type: string
        recommendations:
          type: array
          items:
            type: string
        recoveryActions:
          type: array
          items:
            type: string
        motivationalMessage:
          type: string
        productivityScore:
          type: number