# lib/api-zod/src/generated/types/breakdownRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Deadline Guardian AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BreakdownRequest {
  taskId: number;
  title: string;
  description?: string;
  estimatedHours?: number;
  deadline?: string;
}