# lib/api-zod/src/generated/types/breakdownResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Deadline Guardian AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Subtask } from './subtask';

export interface BreakdownResult {
  taskId: number;
  subtasks: Subtask[];
}