# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Deadline Guardian AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from './task';

export interface DashboardSummary {
  totalTasks: number;
  completedTasks: number;
  highRiskTasks: number;
  productivityScore: number;
  upcomingDeadlines: Task[];
  recentActivity: Task[];
  completionRate?: number;
  averageRisk?: number;
}