# lib/api-zod/src/generated/types/riskResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Deadline Guardian AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskResultRiskLevel } from './riskResultRiskLevel';

export interface RiskResult {
  taskId: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  riskLevel: RiskResultRiskLevel;
  explanation: string;
  recommendations: string[];
}