# lib/api-zod/src/generated/types/subtask.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Deadline Guardian AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubtaskDifficulty } from './subtaskDifficulty';

export interface Subtask {
  title: string;
  estimatedTime: string;
  difficulty: SubtaskDifficulty;
  dependencies?: string[];
  completed?: boolean;
}