# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Deadline Guardian AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskEnergyLevel } from './taskEnergyLevel';
import type { TaskPriority } from './taskPriority';
import type { TaskRiskLevel } from './taskRiskLevel';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  deadline: string;
  /** @nullable */
  estimatedHours?: number | null;
  /** @nullable */
  availableHours?: number | null;
  /** @nullable */
  energyLevel?: TaskEnergyLevel;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  priority: TaskPriority;
  status: TaskStatus;
  /** @nullable */
  riskScore?: number | null;
  /** @nullable */
  riskLevel?: TaskRiskLevel;
  /** @nullable */
  subtasks?: string | null;
  /** @nullable */
  riskAnalysis?: string | null;
  /** @nullable */
  rescuePlan?: string | null;
  createdAt: string;
  updatedAt: string;
}