# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Deadline Guardian AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputEnergyLevel } from './taskInputEnergyLevel';
import type { TaskInputPriority } from './taskInputPriority';
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  deadline: string;
  estimatedHours?: number;
  availableHours?: number;
  energyLevel?: TaskInputEnergyLevel;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  priority: TaskInputPriority;
  status?: TaskInputStatus;
}